leech: attack vector #4 — narrow distinct/counts/nz tables int64 → int8
The decoder's per-class metadata arrays held Leech-lattice coordinates
(|v| ≤ 32) and multiplicity counts (≤ 24) as int64_t simply because the
codegen defaulted to numpy int64. Every value fits in 8 bits — signed
for distincts/nz, unsigned for counts. Narrowing them shrinks the hot
table footprint 8× and frees L1/L2 cache for codewords_flat (which
still dominates at 11.9 MB and is the next target).

Arrays narrowed:
  f0_distinct_flat        int64 → int8
  f0_counts_flat          int64 → uint8
  f1_distinct_flat        int64 → int8
  f1_counts_flat          int64 → uint8
  multiset_distinct_flat  int64 → int8
  multiset_counts_flat    int64 → uint8
  nz_distinct_desc_flat   int64 → int8

Offset arrays (f0_distinct_ofs, multi_ofs, nz_ofs, etc.) stay int64 —
they index into the ragged storage and routinely exceed int32 ranges
at ms_max=18.

Codegen: antsquant/tools/gen_leech_tables.py now applies a range-
checked astype(int8 / uint8) before _emit_array. The check raises if
any value falls outside the narrower range — guarantees no silent
truncation across future fixture rebuilds.

Kernel side:
  - mistralrs-quant/kernels/leech/leech_decode.cu: d_f0_dist_flat /
    d_f0_cnt_flat / ... __device__ declarations match the new types.
    sizeof() / sizeof() count expressions updated.
  - mistralrs-quant/kernels/leech/leech_multiset_unrank.cuh:
    unrank_multiset signature becomes
      const int8_t* dist_vals, const uint8_t* counts
    Inner `out[i] = static_cast<int8_t>(dist_vals[j])` simplifies to
    `out[i] = dist_vals[j]` (no widening then narrowing roundtrip).

Validation (RTX 5090 sm_120):
  Phase 3 exhaustive:       248/248 OK (287,407,104 blocks bit-equal)
  GEMV correctness:         bit-exact (max_abs_diff = 0)
  Parity-perm equivalence:  bit-exact

Bench (mlp.up_proj 12288x4080, M=1):
  Phase 3 decode-only:   2283.2 µs → 1911.3 µs  (1.19x)
                         915 M → 1.09 G blocks/sec
  GEMV (no parity_perm): 2548.9 µs → 2377.5 µs  (1.07x)
  GEMV (parity_perm):    2548.9 µs → 1774.8 µs  (1.44x combined
                                                  with vector #1)

The decode-only speedup (19%) is larger than GEMV (7%) because GEMV
spends a fixed cost in the dot-product epilogue that doesn't shrink
when tables shrink.

Next: codewords_flat at 11.9 MB dwarfs everything else and is the
single remaining table-bandwidth bottleneck. The LLVQ paper does not
bake codewords — they're regenerated on-the-fly via the Golay
encoder from the leader + r-rank. That's a separate refactor.

+++ mistralrs-quant/kernels/leech/leech_decode.cu
@@ -56,16 +56,20 @@ __constant__ uint8_t  c_T[ltab::N_CLASSES];
 // ──────────────────────────────────────────────────────────────────────────
 __device__ uint32_t d_codewords_flat[sizeof(ltab::codewords_flat) / sizeof(uint32_t)];
 __device__ int64_t  d_codewords_ofs [sizeof(ltab::codewords_ofs)  / sizeof(int64_t)];
-__device__ int64_t  d_f0_dist_flat  [sizeof(ltab::f0_distinct_flat) / sizeof(int64_t)];
+// Attack vector #4: distinct/counts/nz arrays narrowed to 8-bit. Values are
+// Leech-lattice coords (|v| ≤ 32) and counts (≤ 24), so they fit in int8/uint8.
+// 8x smaller storage → 8x more L1/L2 cache fit. The _ofs arrays stay int64
+// because they index into millions of elements.
+__device__ int8_t   d_f0_dist_flat  [sizeof(ltab::f0_distinct_flat) / sizeof(int8_t)];
 __device__ int64_t  d_f0_dist_ofs   [sizeof(ltab::f0_distinct_ofs)  / sizeof(int64_t)];
-__device__ int64_t  d_f0_cnt_flat   [sizeof(ltab::f0_counts_flat)   / sizeof(int64_t)];
-__device__ int64_t  d_f1_dist_flat  [sizeof(ltab::f1_distinct_flat) / sizeof(int64_t)];
+__device__ uint8_t  d_f0_cnt_flat   [sizeof(ltab::f0_counts_flat)   / sizeof(uint8_t)];
+__device__ int8_t   d_f1_dist_flat  [sizeof(ltab::f1_distinct_flat) / sizeof(int8_t)];
 __device__ int64_t  d_f1_dist_ofs   [sizeof(ltab::f1_distinct_ofs)  / sizeof(int64_t)];
-__device__ int64_t  d_f1_cnt_flat   [sizeof(ltab::f1_counts_flat)   / sizeof(int64_t)];
-__device__ int64_t  d_multi_dist_flat[sizeof(ltab::multiset_distinct_flat)/sizeof(int64_t)];
+__device__ uint8_t  d_f1_cnt_flat   [sizeof(ltab::f1_counts_flat)   / sizeof(uint8_t)];
+__device__ int8_t   d_multi_dist_flat[sizeof(ltab::multiset_distinct_flat)/sizeof(int8_t)];
 __device__ int64_t  d_multi_ofs     [sizeof(ltab::multi_ofs)         / sizeof(int64_t)];
-__device__ int64_t  d_multi_cnt_flat[sizeof(ltab::multiset_counts_flat)/sizeof(int64_t)];
-__device__ int64_t  d_nz_flat       [sizeof(ltab::nz_distinct_desc_flat)/sizeof(int64_t)];
+__device__ uint8_t  d_multi_cnt_flat[sizeof(ltab::multiset_counts_flat)/sizeof(uint8_t)];
+__device__ int8_t   d_nz_flat       [sizeof(ltab::nz_distinct_desc_flat)/sizeof(int8_t)];
 __device__ int64_t  d_nz_ofs        [sizeof(ltab::nz_ofs)            / sizeof(int64_t)];
 
 // One-shot host-side initializer. The host caller copies the constexpr data
